Create Appointment

+++ composeApp/src/commonMain/kotlin/domain/bookings/BookingRepository.kt
@@ -10,7 +10,7 @@ interface BookingRepository {
     suspend fun createPendingBookingAppointment(userId: Long, vendorId: Long, serviceId: Long, serviceTypeId: Long, therapistId: Long,
                                                 appointmentTime: Int, day: Int, month: Int, year: Int, serviceLocation: String, serviceStatus: String,
                                                 paymentAmount: Double, paymentMethod: String, bookingStatus: String): Single<PendingBookingAppointmentResponse>
-    suspend fun createAppointment(userId: Long, vendorId: Long, paymentAmount: Double, paymentMethod: String, bookingStatus: String, day: Int, month: Int, year: Int) : Single<ServerResponse>
+    suspend fun createAppointment(userId: Long, vendorId: Long, paymentAmount: Int, paymentMethod: String, bookingStatus: String, day: Int, month: Int, year: Int) : Single<ServerResponse>
     suspend fun getPendingBookingAppointment(userId: Long, bookingStatus: String): Single<PendingBookingAppointmentResponse>
     suspend fun deletePendingBookingAppointment(pendingAppointmentId: Long) : Single<ServerResponse>
 }
