+++ composeApp/src/commonMain/kotlin/domain/bookings/BookingRequestContract.kt
@@ -31,7 +31,7 @@ data class CreateAppointmentRequest(@SerialName("user_id") val userId: Long,
                                            @SerialName("month") val month: Int,
                                            @SerialName("year") val year: Int,
                                            @SerialName("bookingStatus") val bookingStatus: String,
-                                           @SerialName("paymentAmount") val paymentAmount: Double,
+                                           @SerialName("paymentAmount") val paymentAmount: Int,
                                            @SerialName("paymentMethod") val paymentMethod: String)
 
 @Serializable
